+++ front/src/app/components/immersion-offer/useTranscientDataFromStorage.ts
@@ -1,17 +1,9 @@
-import ButtonsGroup from "@codegouvfr/react-dsfr/ButtonsGroup";
-import { createModal } from "@codegouvfr/react-dsfr/Modal";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { keys } from "shared";
 import { ContactEstablishmentByMailDto, OmitFromExistingKeys } from "shared";
-import { inputsLabelsByKey } from "src/app/components/immersion-offer/ContactByEmail";
 
 const transcientDataStorageKey = "IfTranscientData";
 const preferUseTranscientDataStorageKey = "IfPreferUseTranscientData";
-
-const transcientPreferencesModal = createModal({
-  id: "transcient-preferences-modal",
-  isOpenedByDefault: true,
-});
 
 const unrelevantDataKeysForContactScope: (keyof ContactEstablishmentByMailDto)[] =
   ["locationId", "message", "appellationCode", "contactMode", "siret"] as const;
@@ -26,7 +18,9 @@ type TranscientDataItem<T> = {
   expireAt: number;
 };
 
-type TranscientData = {
+export const transcientExpirationTimeInMinutes = 10;
+
+export type TranscientData = {
   "contact-establishment": TranscientDataItem<ContactTranscientData>;
 };
 
@@ -44,7 +38,10 @@ const isStringTranscientData = (data: object): data is TranscientData => {
   return data && typeof data === "object" && "contact-establishment" in data;
 };
 
-export const useTranscientDataFromStorage = (scope: keyof TranscientData) => {
+export const useTranscientDataFromStorage = (
+  scope: keyof TranscientData,
+  keepPreferencesInStorage = true,
+) => {
   const dataInStorage = localStorage.getItem(transcientDataStorageKey);
   const preferUseTranscientDataInStorage = localStorage.getItem(
     preferUseTranscientDataStorageKey,
@@ -55,7 +52,7 @@ export const useTranscientDataFromStorage = (scope: keyof TranscientData) => {
   );
   const [preferUseTranscientData, setPreferUseTranscientData] =
     useState<PreferTranscientData>(
-      preferUseTranscientDataInStorage
+      keepPreferencesInStorage && preferUseTranscientDataInStorage
         ? JSON.parse(preferUseTranscientDataInStorage)
         : null,
     );
@@ -66,10 +63,12 @@ export const useTranscientDataFromStorage = (scope: keyof TranscientData) => {
       [scope]: preferUse,
     };
     setPreferUseTranscientData(updatedData);
-    localStorage.setItem(
-      preferUseTranscientDataStorageKey,
-      JSON.stringify(updatedData),
-    );
+    if (keepPreferencesInStorage) {
+      localStorage.setItem(
+        preferUseTranscientDataStorageKey,
+        JSON.stringify(updatedData),
+      );
+    }
   };
 
   const getExpireAt = (expireInMinutes: number) => {
@@ -108,7 +107,9 @@ export const useTranscientDataFromStorage = (scope: keyof TranscientData) => {
       const shouldDataExpire = data?.expireAt && data.expireAt < Date.now();
       if (shouldDataExpire) {
         localStorage.removeItem(transcientDataStorageKey);
-        localStorage.removeItem(preferUseTranscientDataStorageKey);
+        if (keepPreferencesInStorage) {
+          localStorage.removeItem(preferUseTranscientDataStorageKey);
+        }
         return {
           expireAt: 0,
           value: null,
@@ -126,79 +127,5 @@ export const useTranscientDataFromStorage = (scope: keyof TranscientData) => {
     setTranscientData,
     getTranscientDataForScope,
     setTranscientDataForScope,
-    transcientModalPreferences: {
-      ...transcientPreferencesModal,
-      Component: TranscientPreferencesModal,
-    },
   };
-};
-
-const TranscientPreferencesModal = ({
-  scope,
-  onPreferencesChange,
-}: {
-  scope: keyof TranscientData;
-  onPreferencesChange: (accept: boolean) => void;
-}) => {
-  const {
-    getTranscientDataForScope,
-    getPreferUseTranscientDataForScope,
-    setPreferUseTranscientDataForScope,
-    setTranscientDataForScope,
-  } = useTranscientDataFromStorage(scope);
-  const transcientDataForScope = getTranscientDataForScope();
-  const preferUseTranscientData = getPreferUseTranscientDataForScope();
-  const savePreferences = (accept: boolean) => {
-    setPreferUseTranscientDataForScope(accept);
-    setTranscientDataForScope({
-      ...transcientDataForScope?.value,
-    });
-    onPreferencesChange(accept);
-    transcientPreferencesModal.close();
-  };
-  const shouldOpenModal =
-    preferUseTranscientData === null && transcientDataForScope !== null;
-  useEffect(() => {
-    if (shouldOpenModal) {
-      transcientPreferencesModal.open();
-    }
-  }, [shouldOpenModal]);
-
-  if (preferUseTranscientData !== null) {
-    return null;
-  }
-  return (
-    <transcientPreferencesModal.Component title="Préremplir le formulaire">
-      <p>
-        Nous avons trouvé de la données précédemment saisie pour ce formulaire :
-      </p>
-      <ul>
-        {keys(transcientDataForScope?.value ?? {}).map((key) => (
-          <li key={key}>
-            {inputsLabelsByKey[key]}: {transcientDataForScope?.value?.[key]}
-          </li>
-        ))}
-      </ul>
-      <p>Voulez-vous utiliser ces données ?</p>
-      <ButtonsGroup
-        buttons={[
-          {
-            type: "button",
-            children: "Oui",
-            onClick: () => {
-              savePreferences(true);
-            },
-          },
-          {
-            type: "button",
-            priority: "secondary",
-            children: "Non",
-            onClick: () => {
-              savePreferences(false);
-            },
-          },
-        ]}
-      />
-    </transcientPreferencesModal.Component>
-  );
 };
